fusekit: fix type mismatch

+++ fs/fusekit/util.go
@@ -29,28 +29,28 @@ func applyStat(out *fuse.Attr, fi iofs.FileInfo) {
 
 func openFlags(flags uint32) []string {
 	var flagStrs []string
-	if flags&os.O_RDONLY != 0 {
+	if flags&uint32(os.O_RDONLY) != 0 {
 		flagStrs = append(flagStrs, "O_RDONLY")
 	}
-	if flags&os.O_WRONLY != 0 {
+	if flags&uint32(os.O_WRONLY) != 0 {
 		flagStrs = append(flagStrs, "O_WRONLY")
 	}
-	if flags&os.O_RDWR != 0 {
+	if flags&uint32(os.O_RDWR) != 0 {
 		flagStrs = append(flagStrs, "O_RDWR")
 	}
-	if flags&os.O_APPEND != 0 {
+	if flags&uint32(os.O_APPEND) != 0 {
 		flagStrs = append(flagStrs, "O_APPEND")
 	}
-	if flags&os.O_CREATE != 0 {
+	if flags&uint32(os.O_CREATE) != 0 {
 		flagStrs = append(flagStrs, "O_CREAT")
 	}
-	if flags&os.O_EXCL != 0 {
+	if flags&uint32(os.O_EXCL) != 0 {
 		flagStrs = append(flagStrs, "O_EXCL")
 	}
-	if flags&os.O_SYNC != 0 {
+	if flags&uint32(os.O_SYNC) != 0 {
 		flagStrs = append(flagStrs, "O_SYNC")
 	}
-	if flags&os.O_TRUNC != 0 {
+	if flags&uint32(os.O_TRUNC) != 0 {
 		flagStrs = append(flagStrs, "O_TRUNC")
 	}
 	return flagStrs
